+++ src/providers/colors/ColorProvider.cpp
@@ -119,6 +119,19 @@ void ColorProvider::initTypeColorMap()
              std::make_shared<QColor>(
                  HighlightPhrase::FALLBACK_REDEEMED_HIGHLIGHT_COLOR)});
     }
+
+    customColor = getSettings()->liveHighlightColor;
+    if (QColor(customColor).isValid())
+    {
+        this->typeColorMap_.insert(
+            {ColorType::Live, std::make_shared<QColor>(customColor)});
+    }
+    else
+    {
+        this->typeColorMap_.insert(
+            {ColorType::Live, std::make_shared<QColor>(
+                                  HighlightPhrase::FALLBACK_HIGHLIGHT_COLOR)});
+    }
 }
 
 void ColorProvider::initDefaultColors()