+++ src/providers/colors/ColorProvider.hpp
@@ -11,6 +11,7 @@ enum class ColorType {
     SelfHighlight,
     Subscription,
     Whisper,
+    Live,
     RedeemedHighlight
 };